hotel-management-using-django-master/myproject/requirements.txt
asgiref==3.5.2
distlib==0.3.4
Django==4.0.5
django-bootstrap3==21.2
django-crispy-forms==1.14.0
django-pagedown==2.2.1
django-phonenumber-field==6.3.0
django-widget-tweaks==1.4.12
filelock==3.7.1
phonenumbers==8.12.50
Pillow==9.1.1
platformdirs==2.5.2
pygame==2.1.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.1
